refactor(validation): split pipeline into individual gate modules
Extracted all 9 validation gates into separate modules under worker/validation/gates/ to improve testability and maintainability.
- Created worker/validation/types.ts for shared validation types and helpers.
- Created worker/validation/pipeline.ts as the new orchestrator.
- Added individual unit tests for each gate in tests/unit/gates/.
- Updated all call sites and imports to use the new validation path.
- Verified all files are well below the 500-line limit.

Fixes # (no issue number provided)
